+++ Restaurant Day/Views/UIView+Extras.m
@@ -12,42 +12,42 @@
 
 @dynamic x, y, width, height;
 
-- (int)x
+- (NSInteger)x
 {
     return self.frame.origin.x;
 }
 
-- (void)setX:(int)x
+- (void)setX:(NSInteger)x
 {
     self.frame = CGRectMake(x, self.y, self.width, self.height);
 }
 
-- (int)y
+- (NSInteger)y
 {
     return self.frame.origin.y;
 }
 
-- (void)setY:(int)y
+- (void)setY:(NSInteger)y
 {
     self.frame = CGRectMake(self.x, y, self.width, self.height);
 }
 
-- (int)width
+- (NSInteger)width
 {
     return self.frame.size.width;
 }
 
-- (void)setWidth:(int)width
+- (void)setWidth:(NSInteger)width
 {
     self.frame = CGRectMake(self.x, self.y, width, self.height);
 }
 
-- (int)height
+- (NSInteger)height
 {
     return self.frame.size.height;
 }
 
-- (void)setHeight:(int)height
+- (void)setHeight:(NSInteger)height
 {
     self.frame = CGRectMake(self.x, self.y, self.width, height);
 }
